requirements.txt
Django
google-cloud-datastore==1.10.0
google-auth==1.11.2
requests==2.22.0
networkx
google-cloud-storage==1.24.1
googleapis_common_protos
dj_database_url
